Wire hermes-agent's chat_with_hermes into dev/tools/test-unit
test_hermes.py wasn't actually a unit test: it shelled out to kubectl to
find a live hermes-agent pod, started a real kubectl port-forward, and hit
its HTTP endpoint, using sys.exit() instead of assertions. Wiring it into
test-unit as-is would hang or fail immediately in every CI run, since
there's no cluster in a plain unit-test job.

Renamed it to tester.py (matching python-runtime-sandbox's naming for the
same kind of manual/live-pod smoke script -- out of pytest's test_*.py
discovery pattern on purpose, so it keeps its manual utility without being
collected) and wrote a real test_hermes.py that mocks
urllib.request.urlopen to cover chat_with_hermes()'s three branches:
success, non-200 status, and connection failure. Verified passing in an
isolated venv before wiring into PYTHON_TEST_SUITES.
